Add Java and Python error diffusion homework projects
Initial commit of Homework #4 for error diffusion in both Java and Python. Includes Java source code for sequential and parallel error diffusion algorithms, benchmarking, and performance charting, as well as Python scripts and a variety of test images and outputs for both implementations. Project and IDE configuration files are also included.